+++ crates/zng-app/src/view_process.rs
@@ -17,7 +17,7 @@ use crate::{
 
 use parking_lot::{MappedRwLockReadGuard, MappedRwLockWriteGuard, RwLock};
 use zng_app_context::app_local;
-use zng_layout::unit::{DipPoint, DipRect, DipSize, Factor, Px, PxPoint, PxRect, PxSize};
+use zng_layout::unit::{DipPoint, DipRect, DipSideOffsets, DipSize, Factor, Px, PxPoint, PxRect, PxSize};
 use zng_task::SignalOnce;
 use zng_txt::Txt;
 use zng_var::ResponderVar;
@@ -650,6 +650,13 @@ pub struct WindowOpenData {
 
     /// Actual render mode, can be different from the requested mode if it is not available.
     pub render_mode: RenderMode,
+
+    /// Padding that must be applied to the window content so that it stays clear of screen obstructions
+    /// such as a camera notch cutout.
+    ///
+    /// Note that the *unsafe* area must still be rendered as it may be partially visible, just don't place nay
+    /// interactive or important content outside of this padding.
+    pub safe_padding: DipSideOffsets,
 }
 impl WindowOpenData {
     pub(crate) fn new(data: zng_view_api::window::WindowOpenData, map_monitor: impl FnOnce(ApiMonitorId) -> MonitorId) -> Self {
@@ -660,6 +667,7 @@ impl WindowOpenData {
             size: data.size,
             scale_factor: data.scale_factor,
             render_mode: data.render_mode,
+            safe_padding: data.safe_padding,
         }
     }
 }